requirements.txt
fastapi
uvicorn
requests
pydantic
sentence-transformers
streamlit
python-dotenv
torch
